requirements.txt
runpod
huggingface_hub
Pillow
